extract create-team-modal component
