+++ kits/Livewire/Teams/Base/resources/views/layouts/app/header.blade.php
@@ -79,6 +79,8 @@
 
         {{ $slot }}
 
+        <livewire:create-team-modal />
+
         @persist('toast')
             <flux:toast.group>
                 <flux:toast />
