+++ kits/Livewire/Teams/Base/resources/views/layouts/app/sidebar.blade.php
@@ -92,6 +92,8 @@
 
         {{ $slot }}
 
+        <livewire:create-team-modal />
+
         @persist('toast')
             <flux:toast.group>
                 <flux:toast />
